tests/wpt/web-platform-tests/css/css-contain/container-queries/container-type-computed.html

<!doctype html>
<meta charset="utf-8">
<title>Computed values of container-type</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-type">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="support/cq-testcommon.js"></script>
<div id="target"></div>
<script>
setup(() => assert_implements_container_queries());

test_computed_value('container-type', 'initial', 'none');
test_computed_value('container-type', 'unset', 'none');
test_computed_value('container-type', 'inline-size');
test_computed_value('container-type', 'size');
</script>
